requirements.txt
python-telegram-bot==20.8
httpx==0.27.0
nest_asyncio==1.6.0

